Removed abjad.Glissando spanner; use abjad.glissando() instead. Tag #1033.
REMOVED:

    * abjad.Glissando

NEW:

    * abjad.GlissandoIndicator
