+++ abjad/lilypondnames/__init__.py
@@ -12,3 +12,4 @@ from .LilyPondGrobNameManager import LilyPondGrobNameManager
 from .LilyPondGrobOverride import LilyPondGrobOverride
 from .LilyPondSettingNameManager import LilyPondSettingNameManager
 from .LilyPondTweakManager import LilyPondTweakManager
+from .LilyPondTweakManager import TweaksTyping
