+++ abjad/spanners/__init__.py
@@ -4,12 +4,12 @@ Tools for modeling spanners: slurs, beams, glissandi, etc.
 
 from .Spanner import Spanner
 from .Spanner import beam
+from .Spanner import glissando
 from .Spanner import hairpin
 from .Spanner import phrasing_slur
 from .Spanner import slur
 from .Spanner import text_spanner
 from .BowContactSpanner import BowContactSpanner
-from .Glissando import Glissando
 from .HorizontalBracket import HorizontalBracket
 from .OctavationSpanner import OctavationSpanner
 from .PianoPedalSpanner import PianoPedalSpanner
