--- tests/test_Parentage__get_spanners_that_dominate_component_pair.py
@@ -1,191 +1,0 @@
-import abjad
-
-
-def test_Parentage__get_spanners_that_dominate_component_pair_01():
-    """
-    Returns Python list of (spanner, index) pairs.
-    Each spanner dominates an empty slice between components.
-    No spanners dominate voice[0:0].
-    """
-
-    voice = abjad.Voice("{ c'8 d'8 } { e'8 f'8 } { g'8 a'8 }")
-    leaves = abjad.select(voice).leaves()
-    abjad.beam(leaves[:4])
-    glissando = abjad.Glissando()
-    abjad.attach(glissando, leaves[-4:])
-    trill = abjad.TrillSpanner()
-    abjad.attach(trill, leaves)
-
-    assert format(voice) == abjad.String.normalize(
-        r"""
-        \new Voice
-        {
-            {
-                c'8
-                [
-                \startTrillSpan
-                d'8
-            }
-            {
-                e'8
-                \glissando
-                f'8
-                ]
-                \glissando
-            }
-            {
-                g'8
-                \glissando
-                a'8
-                \stopTrillSpan
-            }
-        }
-        """
-        )
-
-    receipt = voice._get_spanners_that_dominate_component_pair(None, voice[0])
-
-    assert len(receipt) == 0
-    assert receipt == []
-
-
-def test_Parentage__get_spanners_that_dominate_component_pair_02():
-    """
-    Beam and trill both dominate crack at voice[1:1].
-    """
-
-    voice = abjad.Voice("{ c'8 d'8 } { e'8 f'8 } { g'8 a'8 }")
-    leaves = abjad.select(voice).leaves()
-    abjad.beam(leaves[:4])
-    glissando = abjad.Glissando()
-    abjad.attach(glissando, leaves[-4:])
-    trill = abjad.TrillSpanner()
-    abjad.attach(trill, leaves)
-
-    assert format(voice) == abjad.String.normalize(
-        r"""
-        \new Voice
-        {
-            {
-                c'8
-                [
-                \startTrillSpan
-                d'8
-            }
-            {
-                e'8
-                \glissando
-                f'8
-                ]
-                \glissando
-            }
-            {
-                g'8
-                \glissando
-                a'8
-                \stopTrillSpan
-            }
-        }
-        """
-        )
-
-    pair = (voice[0], voice[1])
-    receipt = voice._get_spanners_that_dominate_component_pair(*pair)
-
-    assert len(receipt) == 1
-    assert (trill, 2) in receipt
-
-
-def test_Parentage__get_spanners_that_dominate_component_pair_03():
-    """
-    Glissando and trill both dominate crack at voice[2:2].
-    """
-
-    voice = abjad.Voice("{ c'8 d'8 } { e'8 f'8 } { g'8 a'8 }")
-    leaves = abjad.select(voice).leaves()
-    abjad.beam(leaves[:4])
-    glissando = abjad.Glissando()
-    abjad.attach(glissando, leaves[-4:])
-    trill = abjad.TrillSpanner()
-    abjad.attach(trill, leaves)
-
-    assert format(voice) == abjad.String.normalize(
-        r"""
-        \new Voice
-        {
-            {
-                c'8
-                [
-                \startTrillSpan
-                d'8
-            }
-            {
-                e'8
-                \glissando
-                f'8
-                ]
-                \glissando
-            }
-            {
-                g'8
-                \glissando
-                a'8
-                \stopTrillSpan
-            }
-        }
-        """
-        )
-
-    pair = (voice[1], voice[2])
-    receipt = voice._get_spanners_that_dominate_component_pair(*pair)
-
-    assert len(receipt) == 2
-    assert (glissando, 2) in receipt
-    assert (trill, 4) in receipt
-
-
-def test_Parentage__get_spanners_that_dominate_component_pair_04():
-    """
-    No spanners dominate empty slice following voice.
-    """
-
-    voice = abjad.Voice("{ c'8 d'8 } { e'8 f'8 } { g'8 a'8 }")
-    leaves = abjad.select(voice).leaves()
-    abjad.beam(leaves[:4])
-    glissando = abjad.Glissando()
-    abjad.attach(glissando, leaves[-4:])
-    trill = abjad.TrillSpanner()
-    abjad.attach(trill, leaves)
-
-    assert format(voice) == abjad.String.normalize(
-        r"""
-        \new Voice
-        {
-            {
-                c'8
-                [
-                \startTrillSpan
-                d'8
-            }
-            {
-                e'8
-                \glissando
-                f'8
-                ]
-                \glissando
-            }
-            {
-                g'8
-                \glissando
-                a'8
-                \stopTrillSpan
-            }
-        }
-        """
-        )
-
-    pair = (voice[2], None)
-    receipt = voice._get_spanners_that_dominate_component_pair(*pair)
-
-    assert len(receipt) == 0
-    assert receipt == []
